move edit visit components into its own file
